refactor: Backend code refactor
