--- backend/classes/__init__.py
@@ -1,5 +1,0 @@
-from backend.classes.LandsatAPI import LandsatAPI
-from backend.classes.LandsatData import LandsatData
-
-LandsatAPI = LandsatAPI
-LandsatData = LandsatData
