+++ backend/entities/__init__.py
@@ -1,0 +1,5 @@
+from backend.entities.LandsatAPI import LandsatAPI
+from backend.entities.LandsatData import LandsatData
+
+LandsatAPI = LandsatAPI
+LandsatData = LandsatData
