+++ backend/services/ReportGeneratorService.py
@@ -1,9 +1,9 @@
-from classes import LandSatData
+from backend.entities import LandsatData
 
 
 class ReportGeneratorService:
-    def __init__(self, Data: LandSatData):
-        self.Data = Data
+    def __init__(self, data: LandsatData):
+        self.Data = data
 
     def generate_csv_report(self):
         raise NotImplementedError
